ADD CONFIG.JS AND return an array with results and rating films
