+++ src/config.js
@@ -1,0 +1,9 @@
+const config = {
+    urlTopFilms: 'https://www.imdb.com/chart/top/',
+    urlMostPopular: 'https://www.imdb.com/chart/moviemeter',
+    principalElement: '.lister-list tr',
+    titleClass: '.titleColumn a',
+    rating: '.imdbRating strong'
+}
+
+module.exports = config;
